+++ airbyte-integrations/bases/standard-destination-test/src/main/java/io/airbyte/integrations/standardtest/destination/DestinationAcceptanceTest.java
@@ -43,9 +43,9 @@ import io.airbyte.protocol.models.DestinationSyncMode;
 import io.airbyte.protocol.models.Field;
 import io.airbyte.protocol.models.JsonSchemaType;
 import io.airbyte.protocol.models.SyncMode;
-import io.airbyte.workers.general_workers.DbtTransformationRunner;
-import io.airbyte.workers.general_workers.DefaultCheckConnectionWorker;
-import io.airbyte.workers.general_workers.DefaultGetSpecWorker;
+import io.airbyte.workers.generalworkers.DbtTransformationRunner;
+import io.airbyte.workers.generalworkers.DefaultCheckConnectionWorker;
+import io.airbyte.workers.generalworkers.DefaultGetSpecWorker;
 import io.airbyte.workers.WorkerConfigs;
 import io.airbyte.workers.WorkerException;
 import io.airbyte.workers.normalization.NormalizationRunner;
