+++ airbyte-integrations/bases/standard-source-test/src/main/java/io/airbyte/integrations/standardtest/source/AbstractSourceConnectorTest.java
@@ -22,9 +22,9 @@ import io.airbyte.protocol.models.AirbyteMessage.Type;
 import io.airbyte.protocol.models.AirbyteRecordMessage;
 import io.airbyte.protocol.models.ConfiguredAirbyteCatalog;
 import io.airbyte.protocol.models.ConnectorSpecification;
-import io.airbyte.workers.general_workers.DefaultCheckConnectionWorker;
-import io.airbyte.workers.general_workers.DefaultDiscoverCatalogWorker;
-import io.airbyte.workers.general_workers.DefaultGetSpecWorker;
+import io.airbyte.workers.generalworkers.DefaultCheckConnectionWorker;
+import io.airbyte.workers.generalworkers.DefaultDiscoverCatalogWorker;
+import io.airbyte.workers.generalworkers.DefaultGetSpecWorker;
 import io.airbyte.workers.WorkerConfigs;
 import io.airbyte.workers.WorkerException;
 import io.airbyte.workers.process.AirbyteIntegrationLauncher;
